requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
openai==0.27.0
geopy==2.2.0
pytz==2021.1 